+++ CHANGELOG.md
@@ -6,8 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- Added `DefaultReserver` which works just like `OrderedReserver` with one exception: it does not sort the queues
-- Added `BeforeEnqueue` event support
+- Added the `DefaultReserver` which works just like `OrderedReserver` with one exception: it does not sort the queues
+- Added the `BeforeEnqueue` event support
+- Added the `Qless\Subscribers\WatchdogSubscriber` subscriber to watching events for a job
+- Added the `Qless\SystemFacade` facade to to easily mock system functions
+
+### Changed
+- Qless now use `predis/predis` as an internal Redis client. The `redis` extension no longer required
+
+### Removed
+- Removed no longer used internal classes:
+  - `Qless\Events\QlessCoreEventFactory`
+  - `Qless\Subscribers\QlessCoreSubscriber`
+  - `Qless\Redis`
 
 ### Fixed
 - Fixed  `OrderedReserver` to sort queues using natural ordering like [`natsort()`](http://php.net/manual/en/function.natsort.php)